# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Financial Sandbox API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type FinancialItemFrequency =
  (typeof FinancialItemFrequency)[keyof typeof FinancialItemFrequency];

export const FinancialItemFrequency = {
  monthly: "monthly",
  yearly: "yearly",
} as const;

export interface FinancialItem {
  id: string;
  name: string;
  amount: number;
  frequency: FinancialItemFrequency;
}

export interface AssetItem {
  id: string;
  name: string;
  amount: number;
  returnRate: number;
}

export type InvestmentItemFrequency =
  (typeof InvestmentItemFrequency)[keyof typeof InvestmentItemFrequency];

export const InvestmentItemFrequency = {
  monthly: "monthly",
  yearly: "yearly",
} as const;

export interface InvestmentItem {
  id: string;
  name: string;
  amount: number;
  frequency: InvestmentItemFrequency;
  returnRate: number;
}

/**
 * A named allocation pocket within the total rolling investment pool
 */
export interface InvestmentBucket {
  id: string;
  name: string;
  /** Current percentage allocation of the total investment pool (0-100) */
  currentAllocationPct: number;
  /** Target (what-if) percentage allocation (0-100) */
  targetAllocationPct: number;
  /** Expected annual return rate as a decimal (e.g. 0.08 for 8%) */
  returnRate: number;
}

export type ActiveTradingConfigFrequency =
  (typeof ActiveTradingConfigFrequency)[keyof typeof ActiveTradingConfigFrequency];

export const ActiveTradingConfigFrequency = {
  monthly: "monthly",
  yearly: "yearly",
} as const;

/**
 * Special active trading pocket with its own contribution and return rates
 */
export interface ActiveTradingConfig {
  enabled: boolean;
  /** Contribution amount */
  amount: number;
  frequency: ActiveTradingConfigFrequency;
  /** Current expected annual return (can be negative), as a decimal */
  currentReturnRate: number;
  /** Target (what-if) annual return, as a decimal */
  targetReturnRate: number;
  /**
   * Subjective risk level from 1 (low) to 10 (extreme)
   * @minimum 1
   * @maximum 10
   */
  riskLevel: number;
}

export interface FinancialGoal {
  id: string;
  name: string;
  targetAmount: number;
  /** Relative projection year for the goal */
  targetYear: number;
}

export type ScenarioSettingsCompoundingFrequency =
  (typeof ScenarioSettingsCompoundingFrequency)[keyof typeof ScenarioSettingsCompoundingFrequency];

export const ScenarioSettingsCompoundingFrequency = {
  annual: "annual",
  monthly: "monthly",
  quarterly: "quarterly",
} as const;

export interface ScenarioSettings {
  expenseModifier: number;
  incomeModifier: number;
  /** Override annual return for all investments (-1 = use individual rates) */
  investmentReturnOverride: number;
  inflationRate: number;
  compoundingFrequency: ScenarioSettingsCompoundingFrequency;
  timeframeYears: number;
  /** When true, use targetAllocationPct instead of currentAllocationPct for buckets */
  useTargetAllocations: boolean;
  goals: FinancialGoal[];
}

export interface BalanceSheetRow {
  /** header | line | subtotal */
  rowType: string;
  id?: string;
  name: string;
  amount?: number | null;
  amountAnnualized?: number | null;
  amountScenarioAnnualized?: number | null;
  unit?: string;
  frequency?: string;
  returnRate?: number | null;
  pctCurrent?: number | null;
  pctTarget?: number | null;
  note?: string;
}

export interface BalanceSheetSection {
  id: string;
  title: string;
  rows: BalanceSheetRow[];
}

export interface BalanceSheetResponse {
  generatedAt: string;
  sections: BalanceSheetSection[];
}

export interface FinancialProfile {
  income: FinancialItem[];
  expenses: FinancialItem[];
  assets: AssetItem[];
  investments: InvestmentItem[];
  /** Percentage-split allocation buckets within the investment pool */
  investmentBuckets: InvestmentBucket[];
  activeTrading: ActiveTradingConfig;
  scenarioSettings: ScenarioSettings;
}