# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Financial Sandbox API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: financial
    description: Financial data operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /budget:
    get:
      operationId: getBudget
      tags: [financial]
      summary: Get the user's full financial data
      responses:
        "200":
          description: Full financial profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialProfile"
    put:
      operationId: updateBudget
      tags: [financial]
      summary: Update the user's full financial data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinancialProfile"
      responses:
        "200":
          description: Updated financial profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialProfile"

  /budget/reset:
    post:
      operationId: resetBudget
      tags: [financial]
      summary: Reset to default financial data
      responses:
        "200":
          description: Reset financial profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialProfile"

  /budget/balance-sheet:
    get:
      operationId: getBalanceSheet
      tags: [financial]
      summary: Balance sheet for saved profile (JSON)
      responses:
        "200":
          description: Grouped balance sheet sections
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BalanceSheetResponse"
    post:
      operationId: postBalanceSheet
      tags: [financial]
      summary: Balance sheet from a financial profile payload (JSON)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinancialProfile"
      responses:
        "200":
          description: Grouped balance sheet sections
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BalanceSheetResponse"

  /budget/balance-sheet/export.csv:
    post:
      operationId: postBalanceSheetExportCsv
      tags: [financial]
      summary: Export balance sheet as CSV from profile body
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinancialProfile"
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
                format: binary

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    FinancialItem:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        amount:
          type: number
        frequency:
          type: string
          enum: [monthly, yearly]
      required:
        - id
        - name
        - amount
        - frequency

    AssetItem:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        amount:
          type: number
        returnRate:
          type: number
      required:
        - id
        - name
        - amount
        - returnRate

    InvestmentItem:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        amount:
          type: number
        frequency:
          type: string
          enum: [monthly, yearly]
        returnRate:
          type: number
      required:
        - id
        - name
        - amount
        - frequency
        - returnRate

    InvestmentBucket:
      type: object
      description: A named allocation pocket within the total rolling investment pool
      properties:
        id:
          type: string
        name:
          type: string
        currentAllocationPct:
          type: number
          description: Current percentage allocation of the total investment pool (0-100)
        targetAllocationPct:
          type: number
          description: Target (what-if) percentage allocation (0-100)
        returnRate:
          type: number
          description: Expected annual return rate as a decimal (e.g. 0.08 for 8%)
      required:
        - id
        - name
        - currentAllocationPct
        - targetAllocationPct
        - returnRate

    ActiveTradingConfig:
      type: object
      description: Special active trading pocket with its own contribution and return rates
      properties:
        enabled:
          type: boolean
        amount:
          type: number
          description: Contribution amount
        frequency:
          type: string
          enum: [monthly, yearly]
        currentReturnRate:
          type: number
          description: Current expected annual return (can be negative), as a decimal
        targetReturnRate:
          type: number
          description: Target (what-if) annual return, as a decimal
        riskLevel:
          type: integer
          minimum: 1
          maximum: 10
          description: Subjective risk level from 1 (low) to 10 (extreme)
      required:
        - enabled
        - amount
        - frequency
        - currentReturnRate
        - targetReturnRate
        - riskLevel

    FinancialGoal:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        targetAmount:
          type: number
        targetYear:
          type: number
          description: Relative projection year for the goal
      required:
        - id
        - name
        - targetAmount
        - targetYear

    ScenarioSettings:
      type: object
      properties:
        expenseModifier:
          type: number
        incomeModifier:
          type: number
        investmentReturnOverride:
          type: number
          description: Override annual return for all investments (-1 = use individual rates)
        inflationRate:
          type: number
        compoundingFrequency:
          type: string
          enum: [annual, monthly, quarterly]
        timeframeYears:
          type: integer
        useTargetAllocations:
          type: boolean
          description: When true, use targetAllocationPct instead of currentAllocationPct for buckets
        goals:
          type: array
          items:
            $ref: "#/components/schemas/FinancialGoal"
      required:
        - expenseModifier
        - incomeModifier
        - investmentReturnOverride
        - inflationRate
        - compoundingFrequency
        - timeframeYears
        - useTargetAllocations
        - goals

    BalanceSheetRow:
      type: object
      required:
        - rowType
        - name
      properties:
        rowType:
          type: string
          description: header | line | subtotal
        id:
          type: string
        name:
          type: string
        amount:
          type: number
          nullable: true
        amountAnnualized:
          type: number
          nullable: true
        amountScenarioAnnualized:
          type: number
          nullable: true
        unit:
          type: string
        frequency:
          type: string
        returnRate:
          type: number
          nullable: true
        pctCurrent:
          type: number
          nullable: true
        pctTarget:
          type: number
          nullable: true
        note:
          type: string

    BalanceSheetSection:
      type: object
      required:
        - id
        - title
        - rows
      properties:
        id:
          type: string
        title:
          type: string
        rows:
          type: array
          items:
            $ref: "#/components/schemas/BalanceSheetRow"

    BalanceSheetResponse:
      type: object
      required:
        - generatedAt
        - sections
      properties:
        generatedAt:
          type: string
        sections:
          type: array
          items:
            $ref: "#/components/schemas/BalanceSheetSection"

    FinancialProfile:
      type: object
      properties:
        income:
          type: array
          items:
            $ref: "#/components/schemas/FinancialItem"
        expenses:
          type: array
          items:
            $ref: "#/components/schemas/FinancialItem"
        assets:
          type: array
          items:
            $ref: "#/components/schemas/AssetItem"
        investments:
          type: array
          items:
            $ref: "#/components/schemas/InvestmentItem"
        investmentBuckets:
          type: array
          items:
            $ref: "#/components/schemas/InvestmentBucket"
          description: Percentage-split allocation buckets within the investment pool
        activeTrading:
          $ref: "#/components/schemas/ActiveTradingConfig"
        scenarioSettings:
          $ref: "#/components/schemas/ScenarioSettings"
      required:
        - income
        - expenses
        - assets
        - investments
        - investmentBuckets
        - activeTrading
        - scenarioSettings