# lib/api-zod/src/generated/types/balanceSheetResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Financial Sandbox API
 * OpenAPI spec version: 0.1.0
 */
import type { BalanceSheetSection } from "./balanceSheetSection";

export interface BalanceSheetResponse {
  generatedAt: string;
  sections: BalanceSheetSection[];
}