# lib/api-zod/src/generated/types/balanceSheetRow.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Financial Sandbox API
 * OpenAPI spec version: 0.1.0
 */

export interface BalanceSheetRow {
  /** header | line | subtotal */
  rowType: string;
  id?: string;
  name: string;
  amount?: number | null;
  amountAnnualized?: number | null;
  amountScenarioAnnualized?: number | null;
  unit?: string;
  frequency?: string;
  returnRate?: number | null;
  pctCurrent?: number | null;
  pctTarget?: number | null;
  note?: string;
}