# lib/api-zod/src/generated/types/balanceSheetSection.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Financial Sandbox API
 * OpenAPI spec version: 0.1.0
 */
import type { BalanceSheetRow } from "./balanceSheetRow";

export interface BalanceSheetSection {
  id: string;
  title: string;
  rows: BalanceSheetRow[];
}