# lib/api-zod/src/generated/types/financialGoal.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Financial Sandbox API
 * OpenAPI spec version: 0.1.0
 */

export interface FinancialGoal {
  id: string;
  name: string;
  targetAmount: number;
  /** Relative projection year for the goal */
  targetYear: number;
}