# lib/api-zod/src/generated/types/scenarioSettings.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Financial Sandbox API
 * OpenAPI spec version: 0.1.0
 */
import type { FinancialGoal } from "./financialGoal";
import type { ScenarioSettingsCompoundingFrequency } from "./scenarioSettingsCompoundingFrequency";

export interface ScenarioSettings {
  expenseModifier: number;
  incomeModifier: number;
  /** Override annual return for all investments (-1 = use individual rates) */
  investmentReturnOverride: number;
  inflationRate: number;
  compoundingFrequency: ScenarioSettingsCompoundingFrequency;
  timeframeYears: number;
  /** When true, use targetAllocationPct instead of currentAllocationPct for buckets */
  useTargetAllocations: boolean;
  goals: FinancialGoal[];
}